fix(cli): point doctor and status at the real next step (#274)
* fix(agent): point agent doctor at the data dir before a new join

Run without --data-dir, agent doctor looked only at the default data
directory, asserted that the host had never registered, and suggested a
new join string. On a host whose agent runs with --data-dir, following
that hits the name's existing registration, and revoking it cuts the
running agent off. The findings now speak about the directory read, and a
run on the default directory that finds no registered agent.json leads
with rerunning under the agent's --data-dir. The registration's 409
refusal says the same before suggesting a revoke.

A data directory the caller cannot search made agent doctor say that
agent.json "exists but cannot be read" even when it was absent. Such a
run now reads data_dir_unreadable, in the same layer 2 as before.


* fix(cli): name the full rule ID wherever a command is shown

server doctor's next steps embedded the table's elided ID, so running
"wgft rule enable r_01M3AWMEPZ…" as shown failed with not found. The
same held for the deny rm and allow add suggestions and the closing
"traffic stops on rule" line. Those now carry the full ID; tables keep
the short form.

Rule commands also drop a trailing "…" or "..." before prefix matching,
so a short ID pasted from a table works, and an ambiguous prefix now
lists the matching full IDs instead of only saying it matched several.


* fix(server): point a UDP rule's outside test at the real client

server doctor told the operator to test every public port with
nc -vz <this vps> <port>, a TCP connect, so a healthy UDP rule looked
unreachable. A UDP rule's public-port next step and the "from outside"
not-tested item now name the real client, and say that a bare UDP send
such as nc -u cannot show whether the datagram arrived. TCP keeps nc -vz.


* fix(server): name the agent host when its ip_forward stops a rule

A kernel-mode agent whose host has net.ipv4.ip_forward at 0 reports its
LAN rules as "net.ipv4.ip_forward is 0; the kernel does not forward to
a target that is not this host". server doctor filed that under
target_error and told the operator to check the service at the target,
and status repeated it without saying whose host it was; read on the VPS,
"this host" reads as the VPS.

server doctor now classifies the report as agent_ip_forward_off, names
the agent whose host holds the value, and points at that host's sysctl,
an agent restart and agent doctor there. Older agents' wording maps to
the same code. The agent's reason starts with "on the agent host", and
status names the reporting agent for every agent-reported rule error.


* fix(server): report a VPS ip_forward turned off while running

A kernel-mode server writes net.ipv4.ip_forward=1 only at its start. Set
to 0 from outside while it runs, forwarding stopped, yet status said
healthy, server doctor found no failing check, the Web UI said all
systems operational with ip_forward "set by wgft", and server check was
silent when the value was writable.

The server does not write it back: a 0 set while it runs is someone's
explicit act, a sysctl change sends no notice so undoing it would need a
standing poll, and the kernel-mode agent already reports without
reverting. Instead the admin API returns the current value from a
kernel-mode server, server doctor fails the public port of each
kernel-forwarded rule and the dataplane check with ip_forward_off,
status marks the server degraded, the Web UI shows the current value,
and server check reports a writable 0 as well. Proxy rules do not depend
on it and are left as they are.


* fix(agent): fail agent doctor's stream check on a server cert mismatch

After the server was rebuilt with teardown --purge, an agent still
pinning the old certificate made agent doctor report the control
connection as UNKNOWN and say the agent retries by itself, so it was not
a fault. A pin mismatch never clears on retry. stream.connection now
reads FAILED with server_cert_mismatch and gives the same advice as the
agent's log: register again with a new join string, or find what
answers TLS on the path if the server was not rebuilt.

The exit code stays with the tunnel and listener checks, since rules the
agent holds keep forwarding while its tunnel lasts. The agent reports
the mismatch as pin_mismatch on its control socket; older agents are
recognised by their reason text. On the server, the net/http "TLS
handshake error ... bad certificate" line now says it is an agent
pinning another certificate and needs a new join string. It cannot say
which agent: the handshake ends before the agent names itself.


* fix(cli): take an exact rule ID first and require the short form's length

Rule import and the admin API accept any non-empty ID, so an ID can end
in "..." or "…". With rules "web" and "web...", "wgft rule rm web..."
removed "web", and with a single rule "wgft rule rm r_..." removed it.
A rule argument now matches an ID exactly first, mark included, and an
argument whose trailing mark is dropped must keep at least the 12
characters rule ls shows. design.md 10.2 no longer rests on IDs being
ASCII.

Also from the review: the stream.connection paragraph in 10.2c gets back
its tail that had moved under the server log paragraph, 10.2c notes that
--json keeps a top-level ok beside the server_cert_mismatch FAILED, the
help for rule ls and agent doctor is reworded, and tests cover the
unpublished-rule exclusion in IPForwardStop, the doubt on an unregistered
default agent.json, and a generation gap reported together with
ip_forward.


---------

+++ cmd/wgft/doctor.go
@@ -184,7 +184,9 @@ func doctorExit(rep doctorReport) error {
 	var bad []string
 	for _, r := range rep.Rules {
 		if r.Status == statusFailed {
-			bad = append(bad, short(r.RuleID)+" at "+r.StoppedAt)
+			// 終了の 1 行は、次に `server doctor <rule>` へ貼る ID を名指すので、完全な ID を
+			// 使う(設計文書 10.2 節)。
+			bad = append(bad, r.RuleID+" at "+r.StoppedAt)
 		}
 	}
 	if len(bad) == 0 {
